+++ src/main/java/com/github/rayinfinite/scheduler/repository/ClassroomRepository.java
@@ -3,5 +3,5 @@ package com.github.rayinfinite.scheduler.repository;
 import com.github.rayinfinite.scheduler.entity.Classroom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClassroomRepository extends JpaRepository<Classroom, Long> {
+public interface ClassroomRepository extends JpaRepository<Classroom, Integer> {
 }
